Combining BlocklyComp and BlocklyCompEdit
